refactor: api data type 변경
